+++ lib/src/z_base_config/z_base_config.dart
@@ -6,7 +6,8 @@ part 'z_base_config.freezed.dart';
 
 /// An abstract interface that defines the configuration for value parsing and transformation.
 ///
-/// Contains a single flag [nullable] indicating whether `null` is considered a valid input.
+/// Contains a flag [nullable] indicating whether `null` is considered a valid input.
+/// Contains a flag [optional] indicating whether the value can be omitted.
 abstract class Config {
   /// Returns `true` if the configuration allows `null` as a valid input.
   bool get nullable;
@@ -20,7 +21,7 @@ abstract class Config {
 /// This class uses `freezed` to generate equality, copy, and serialization features.
 /// It implements [Config] and stores a list of transformation or rule functions.
 ///
-/// Use methods like [addTransformation], [addRule], or [makeNullable]
+/// Use methods like [addTransformation], [addProcessing], or [makeNullable]
 /// to create updated copies with additional configuration.
 @freezed
 abstract class ZBaseConfig with _$ZBaseConfig implements Config {
@@ -29,23 +30,24 @@ abstract class ZBaseConfig with _$ZBaseConfig implements Config {
   /// - [fns]: A list of transformations or rules to apply.
   /// - [nullable]: Whether `null` is treated as valid input (defaults to `false`).
   const factory ZBaseConfig({
-    required List<TransformAny<dynamic, dynamic>> fns,
+    required List<Operation<dynamic, dynamic>> fns,
     @Default(false) bool nullable,
     @Default(false) bool optional,
   }) = _ZBaseConfig;
 
   const ZBaseConfig._();
 
+  /// Returns a new [ZBaseConfig] with the given operation [o] added to the list.
+  ZBaseConfig _addOperation<From, To>(Operation<From, To> o) => copyWith(fns: [...fns, o]);
+
   /// Returns a new [ZBaseConfig] with the given transformation [t] added to the list.
-  ZBaseConfig addTransformation<From, To>(TransformAny<From, To> t) => copyWith(fns: [...fns, t]);
+  ZBaseConfig addTransformation<From, To>(Transformation<From, To> t) => _addOperation(t);
 
-  /// Returns a new [ZBaseConfig] with the given processor [p] added to the list.
-  ZBaseConfig addProcessor<T>(Processor<T> p) => addTransformation(Process<T>(p));
+  /// Returns a new [ZBaseConfig] with the given processing [p] added to the list.
+  ZBaseConfig addProcessing<T>(Processing<T> p) => _addOperation(p);
 
-  /// Returns a new [ZBaseConfig] with the given validation [r] added as a transformation.
-  ///
-  /// Useful for appending rules without changing the type.
-  ZBaseConfig addRule<T>(RuleAny<T> r) => addTransformation(r);
+  /// Returns a new [ZBaseConfig] with the given validation [v] added to the list.
+  ZBaseConfig addValidation<T>(Validation<T> v) => _addOperation(v);
 
   /// Returns a new [ZBaseConfig] with [nullable] set to `true`.
   ZBaseConfig makeNullable() => copyWith(nullable: true);